Automated processing of CUT&TAG samples

DEPRECATED! Use process_epigenome() with seq_type="CT" instead.

Usage

processCUTTAG(
  fastq_files,
  out_name,
  type = "PE",
  run_fastqc = TRUE,
  path_fastqc = "FastQC/",
  path_bam = "BAM/",
  path_peaks = "Peaks/",
  path_logs = "Logs/",
  index = "/biodata/indices/species/Hsapiens/ucsc.hg19",
  remove = c("chrM", "chrUn", "_random", "_hap", "_gl", "EBV"),
  blacklist = "~/data/consensusBlacklist.bed",
  gen_sizes = "",
  seacr_type = "stingent",
  seacr_top = 0.01,
  cores = 8,
  bedtools_bamtobed = "bedtools bamtobed",
  bedtools_genomecov = "bedtools genomecov",
  seacr = "SEACR_1.3.sh"
)

Arguments

fastq_files

List containing the different FastQ file pairs (one list element for each sample).

out_name

Sample name to use for each of the samples.

type

Character indicating if reads are paired end (PE) or single end (SE).

run_fastqc

Logical indicating whether to run FastQC or not.

path_fastqc

Path for the FastQC results.

path_bam

Path for the aligment and filtering results.

path_peaks

Path for the peak files.

path_logs

Path for the logs.

index

Path to the index needed by Bowtie2.

remove

List of chromosomes to remove from final BAM file.

blacklist

List of blaclisted regions to remove from BAM file.

gen_sizes

Chromosome sizes to use for bedgraph conversion.

seacr_type

Type of peaks to call with SEACR, either "stringent" (default) or "relaxed."

seacr_top

A numeric threshold n between 0 and 1 returns the top n fraction of peaks based on total signal within peaks. Default: 0.01.

cores

Number of threads to use for the analysis.

bedtools_bamtobed

Path or alias of the bedtools bamtobed utility.

bedtools_genomecov

Path or alias of the bedtools genomecov utility.

seacr

Path or alias of the SEACR utilty.